lib: replace: make rep_strerror_r() extern
Bundled heimdal fails to link the rep_strerror_r() symbol as it's
generated as a local instead of global symbol. Declaring the
prototype as 'extern' turns it into a global symbol, which can
be resolved by the linker.

This seems to be related to commit 378ae342c4e2e6b5bafbe54067d35344746be41b